Do not notify the retrying from other streams when free CUDA allocation
